+++ .vscode/httpRequest.code-snippets
@@ -4,9 +4,9 @@
 		"prefix": "hrget",
 		"body": [
 			"HttpRequest({",
-			"    type: 'GET',",
-			"    _url: '$1',",
-			"    _query_string: {",
+			"    method: 'GET',",
+			"    url: '$1',",
+			"    query: {",
 			"        $2",
 			"    },",
 			"    headers: {",
@@ -14,7 +14,7 @@
 			"        Accept: 'application/json, text/plain, */*',",
 			"        Cookie: '$3',",
 			"    },",
-			"    success: chunk => {",
+			"    success: res => {",
 			"        $4",
 			"    },",
 			"    error: err => {",
@@ -29,8 +29,8 @@
 		"prefix": "hrpost",
 		"body": [
 			"HttpRequest({",
-			"    type: 'POST',",
-			"    _url: '$1',",
+			"    method: 'POST',",
+			"    url: '$1',",
 			"    contents: {",
 			"        $2",
 			"    },",
@@ -40,7 +40,7 @@
 			"        'Content-Type': 'application/x-www-form-urlencoded; charset=utf-8',",
 			"        Cookie: '$3',",
 			"    },",
-			"    success: chunk => {",
+			"    success: res => {",
 			"        $4",
 			"    },",
 			"    error: err => {",
