+++ lib/Script.js
@@ -1,0 +1,22 @@
+module.exports = {
+    version: '|version: 3.7.0|in nodejs',
+    author: '@shanmite',
+    UIDs: [
+        213931643,
+        15363359,
+        31252386,
+        80158015,
+        678772444,
+        35719643,
+        223748830,
+        420788931,
+        689949971,
+        38970985
+    ],
+    TAGs: [
+        '抽奖',
+        '互动抽奖',
+        '转发抽奖',
+        '动态抽奖',
+    ]
+};
