+++ main.js
@@ -1,6 +1,25 @@
-import { main,isMe } from './lottery-in-nodejs.js';
-/**@type {string}*/
-const COOKIE = process.env.COOKIE;
-const SCKEY = process.env.SCKEY;
-main(COOKIE);
-isMe(SCKEY);
+const setGlobalVar = require("./lib/setCookie");
+
+const isUndefined = value => typeof value === 'undefined';
+
+const {
+    SCKEY,
+    COOKIE,
+    COOKIE_2,
+    COOKIE_3,
+    COOKIE_4,
+    COOKIE_5,
+} = process.env;
+
+const COOKIE_ARR = [COOKIE, COOKIE_2, COOKIE_3, COOKIE_4, COOKIE_5];
+
+COOKIE_ARR.forEach((cookie, num) => {
+    if (isUndefined(cookie)) return;
+    setGlobalVar(cookie, SCKEY).then(async () => {
+        const { start, isMe, checkCookie } = require("./lib/lottery-in-nodejs");
+        const isRight = await checkCookie(num + 1);
+        if (!isRight) return;
+        isMe();
+        start();
+    })
+});
