+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "description": "run javascript in github",
   "main": "main.js",
-  "type": "module",
   "scripts": {
     "test": "node main.js"
   },
